Workaround for nasty Vivado bug
The bug is present from 2019.2 onwards to 2021.2; we did not test any newer versions yet: It seems that Vivado has a problem with the ALIAS construction used in this component. Without this workaround: When writing to register x"5" and you would expect "timer_a_flag <= '0';": Well: Nothing would happen. With this workaround it seems to work and seems to be stable.