Fixed PLA bugs
When implementing support for actual hardware cartridges for the MEGA65's Expansion Port, we fixed several bugs where this simulation was not in line with the behavior of the real PLA as described in "The C64 PLA Dissected", Revision 1.1 by Thomas 'skoe' Giesel. The Chip Enable / Chip Select signal that came out of fpga64_sid_iec.vhd was wrong in certain situations and also the cs_romL, cs_romH and cs_UMAXromH were buggy. The bugs are only relevant in the context of actual hardware cartridges.