1541 Dual ROM: Fixed Clock Domain Crossing
It was not possible to activate the Dual ROM mode for the simulated 1541 drives without running into timing closure problems. The reason was that while the original MiSTer implementation offered two clock inputs (clk and clk_sys), there was no clock domain crossing implemented. MiSTer works in one clock domain which is why nobody noticed so far, but we work in two: QNICE (registers on falling edge) and the actual core clock.