WIP/research fix Ultimax edge case
Fix incorrect behavior: The /ROMH signal is generated for the $A000-$BFFF range, which is incorrect in Ultimax mode. This fix still has side effects, and is just the start of a research and overall fix of a larger problem: In the current C64 Core on MEGA65 (V5.1), Ultimax mode in Expansion port is incomplete and prevents IDE64 cartridge from functioning properly due to memory mapping limitations. It is not possible to map any external RAM/ROM to $1000-$7FFF in Ultimax mode.

https://github.com/MJoergen/C64MEGA65/issues/176